pyproject.toml
[project]
name = "dev-agent"
version = "1.0.0"
description = "多 Agent 软件开发系统"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.18.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
dev-agent = "src.cli.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "tests*"]
